fix(#932): consolidate useTransactionPolling into useTransaction.ts
- Moved useTransactionPolling hook into useTransaction.ts so all transaction status tracking lives in a single module
- Added architectural comment establishing single-implementation convention
- Updated TransactionStatus.tsx to import from useTransaction.ts
- Deleted useTransactionPolling.ts (duplicate file)
- Downgraded jsdom from 29.0.1 to 25.0.1 to fix ESM compatibility issue blocking all tests
