+++ frontend/package.json
@@ -71,7 +71,7 @@
     "fast-check": "^4.5.3",
     "globals": "^14.0.0",
     "husky": "^9.1.7",
-    "jsdom": "^29.0.1",
+    "jsdom": "^25.0.1",
     "lint-staged": "^16.4.0",
     "postcss": "^8.4.32",
     "prettier": "^3.8.1",
